services/billing/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
stripe==12.0.0
PyJWT==2.8.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.3
